requirements.txt
django-ckeditor==6.4.2
